move temporary workarounds to new package
