+++ plugin/lint-baseline.xml
@@ -1,92 +1,3 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <issues format="6" by="lint 8.10.1" type="baseline" client="gradle" dependencies="false" name="AGP (8.10.1)" variant="all" version="8.10.1">
-
-    <issue
-        id="InternalGradleApiUsage"
-        message="Avoid using internal Gradle APIs"
-        errorLine1="import org.gradle.api.internal.project.ProjectInternal"
-        errorLine2="~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~">
-        <location
-            file="src/main/kotlin/com/vanniktech/maven/publish/Platform.kt"
-            line="8"
-            column="1"/>
-    </issue>
-
-    <issue
-        id="InternalGradleApiUsage"
-        message="Avoid using internal Gradle APIs"
-        errorLine1="import org.gradle.api.plugins.internal.JavaConfigurationVariantMapping"
-        errorLine2="~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~">
-        <location
-            file="src/main/kotlin/com/vanniktech/maven/publish/Platform.kt"
-            line="10"
-            column="1"/>
-    </issue>
-
-    <issue
-        id="InternalGradleApiUsage"
-        message="Avoid using internal Gradle APIs"
-        errorLine1="import org.gradle.api.plugins.internal.JavaPluginHelper"
-        errorLine2="~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~">
-        <location
-            file="src/main/kotlin/com/vanniktech/maven/publish/Platform.kt"
-            line="11"
-            column="1"/>
-    </issue>
-
-    <issue
-        id="InternalGradleApiUsage"
-        message="Avoid using internal Gradle APIs"
-        errorLine1="import org.gradle.api.plugins.internal.JvmPluginsHelper"
-        errorLine2="~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~">
-        <location
-            file="src/main/kotlin/com/vanniktech/maven/publish/Platform.kt"
-            line="12"
-            column="1"/>
-    </issue>
-
-    <issue
-        id="InternalGradleApiUsage"
-        message="Avoid using internal Gradle APIs"
-        errorLine1="import org.gradle.internal.component.external.model.ProjectDerivedCapability"
-        errorLine2="~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~">
-        <location
-            file="src/main/kotlin/com/vanniktech/maven/publish/Platform.kt"
-            line="15"
-            column="1"/>
-    </issue>
-
-    <issue
-        id="InternalGradleApiUsage"
-        message="Avoid using internal Gradle APIs"
-        errorLine1="import org.gradle.jvm.component.internal.DefaultJvmSoftwareComponent"
-        errorLine2="~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~">
-        <location
-            file="src/main/kotlin/com/vanniktech/maven/publish/Platform.kt"
-            line="16"
-            column="1"/>
-    </issue>
-
-    <issue
-        id="InternalGradleApiUsage"
-        message="Avoid using internal Gradle APIs (method createDocumentationVariantWithArtifact from org.gradle.api.plugins.internal.JvmPluginsHelper)"
-        errorLine1="        JvmPluginsHelper.createDocumentationVariantWithArtifact("
-        errorLine2="        ^">
-        <location
-            file="src/main/kotlin/com/vanniktech/maven/publish/Platform.kt"
-            line="502"
-            column="9"/>
-    </issue>
-
-    <issue
-        id="InternalGradleApiUsage"
-        message="Avoid using internal Gradle APIs (method getJavaComponent from org.gradle.api.plugins.internal.JavaPluginHelper)"
-        errorLine1="      val component = JavaPluginHelper.getJavaComponent(project) as DefaultJvmSoftwareComponent"
-        errorLine2="                      ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~">
-        <location
-            file="src/main/kotlin/com/vanniktech/maven/publish/Platform.kt"
-            line="533"
-            column="23"/>
-    </issue>
-
 </issues>
