+++ plugin/src/main/kotlin/com/vanniktech/maven/publish/ProjectExtensions.kt
@@ -7,13 +7,10 @@ import com.android.build.api.variant.AndroidComponentsExtension
 import org.gradle.api.Action
 import org.gradle.api.JavaVersion
 import org.gradle.api.Project
-import org.gradle.api.file.Directory
 import org.gradle.api.plugins.JavaPluginExtension
-import org.gradle.api.provider.Provider
 import org.gradle.api.publish.PublishingExtension
 import org.gradle.api.publish.maven.MavenPublication
 import org.gradle.plugins.signing.SigningExtension
-import org.gradle.util.GradleVersion
 import org.jetbrains.kotlin.gradle.plugin.KotlinBasePlugin
 
 @Suppress(
@@ -87,19 +84,3 @@ internal fun Project.isAtLeastUsingAndroidGradleVersion(major: Int, minor: Int, 
     false
   }
 }
-
-internal fun Project.rootProjectBuildDirCompat(): Provider<Directory> {
-  // TODO: remove this when Gradle 8.8/8.13 is the minimum supported version.
-  return when {
-    GradleVersion.current() >= SETTINGS_DIRECTORY_GRADLE_VERSION -> project.provider {
-      layout.settingsDirectory.dir("build")
-    }
-    GradleVersion.current() >= ISOLATED_PROJECT_VIEW_GRADLE_VERSION -> project.provider {
-      isolated.rootProject.projectDirectory.dir("build")
-    }
-    else -> rootProject.layout.buildDirectory
-  }
-}
-
-private val ISOLATED_PROJECT_VIEW_GRADLE_VERSION = GradleVersion.version("8.8-rc-1")
-private val SETTINGS_DIRECTORY_GRADLE_VERSION = GradleVersion.version("8.13-rc-1")
